Move categoryData.ts to shared location - eliminate cross-feature dependency
WHAT: Move products/utils/categoryData.ts → shared/utils/categoryData.ts

WHY: categoryData (mainCategories) is used by 3 different features:
- Navigation: CategorySidebar.svelte, ShopView.svelte
- Reports: CategoryReportsAdmin.svelte, ReportCategoryDialog.svelte
- Products: ProductBrowserView.svelte, ProductBrowserData.svelte, AllProductsGrid.svelte

This was a clear architectural violation - shared application data
living in a feature-specific location, causing cross-feature imports.

CHANGES:
- Move categoryData.ts from products/utils/ to shared/utils/
- Update 8 import statements across all affected files
- Update products/index.ts export to reference new location
- Establish clean dependency flow: Features → Shared (not Features ↔ Features)
