+++ ui/src/shared/utils/categoryData.ts
